+++ src/components/Table/components/AggregatedCellRenderer/AggregatedCellRenderer.test.tsx
@@ -4,7 +4,7 @@ import { getJestSelectors } from '@volkovlabs/jest-selectors';
 import React from 'react';
 
 import { TEST_IDS } from '@/constants';
-import { CellAggregation, ColumnMeta } from '@/types';
+import { CellAggregation, CellType, ColumnMeta } from '@/types';
 import { createColumnConfig, createField } from '@/utils';
 
 import { AggregatedCellRenderer } from './AggregatedCellRenderer';
@@ -82,5 +82,61 @@ describe('AggregatedCellRenderer', () => {
 
     expect(selectors.root()).toBeInTheDocument();
     expect(selectors.root()).toHaveTextContent('123%');
+  });
+
+  it('Should add contrast color if background passed', async () => {
+    const renderValue: any = jest.fn(() => 123);
+    render(
+      getComponent({
+        column: createColumnWithMeta({
+          config: createColumnConfig({ aggregation: CellAggregation.MIN }),
+          field: createField({ config: { unit: 'percent' } }),
+        }),
+        renderValue: renderValue,
+        bgColor: '#ffc300 ',
+      })
+    );
+
+    expect(selectors.root()).toBeInTheDocument();
+    expect(selectors.root()).toHaveTextContent('123%');
+    expect(selectors.root()).toHaveStyle({
+      color: 'rgb(0, 0, 0)',
+    });
+  });
+
+  it('Should add inherit color if background not specified', async () => {
+    const renderValue: any = jest.fn(() => 123);
+    render(
+      getComponent({
+        column: createColumnWithMeta({
+          config: createColumnConfig({ aggregation: CellAggregation.MIN }),
+          field: createField({ config: { unit: 'percent' } }),
+        }),
+        renderValue: renderValue,
+      })
+    );
+
+    expect(selectors.root()).toBeInTheDocument();
+    expect(selectors.root()).toHaveTextContent('123%');
+    expect(selectors.root().style.color).toBe('');
+  });
+
+  it('Should add text color if config specified', async () => {
+    const renderValue: any = jest.fn(() => 123);
+    render(
+      getComponent({
+        column: createColumnWithMeta({
+          config: createColumnConfig({ type: CellType.COLORED_TEXT, aggregation: CellAggregation.MIN }),
+          field: createField({ config: { unit: 'percent' } }),
+        }),
+        renderValue: renderValue,
+      })
+    );
+
+    expect(selectors.root()).toBeInTheDocument();
+    expect(selectors.root()).toHaveTextContent('123%');
+    expect(selectors.root()).toHaveStyle({
+      color: 'rgb(128, 128, 128)',
+    });
   });
 });